# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plume",
    "version": "0.1.0",
    "description": "API REST de Plume. Toutes les requêtes et réponses sont en JSON (sauf mention). Les requêtes qui modifient un état exigent la session et l’en-tête X-CSRF-Token (jeton renvoyé par /auth/login et /auth/session), ainsi qu’un en-tête Origin identique à l’URL publique."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/folders": {
      "get": {
        "operationId": "getFolders",
        "summary": "Dossiers et compteurs",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/labels": {
      "get": {
        "operationId": "getLabels",
        "summary": "Libellés",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/labels/{name}": {
      "delete": {
        "operationId": "deleteLabelsName",
        "summary": "Supprime un libellé (retiré de tous les messages)",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.+-]{0,63}$"
            }
          }
        ]
      }
    },
    "/messages": {
      "get": {
        "operationId": "getMessages",
        "summary": "Liste paginée des messages",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "default": "INBOX",
              "type": "string",
              "minLength": 1,
              "maxLength": 500
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "default": "all",
              "type": "string",
              "enum": [
                "all",
                "unseen",
                "flagged",
                "attachments"
              ]
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.+-]{0,63}$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ]
      }
    },
    "/messages/{id}": {
      "get": {
        "operationId": "getMessagesId",
        "summary": "Message (en-têtes, pièces jointes)",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      },
      "patch": {
        "operationId": "patchMessagesId",
        "summary": "Lu, suivi, libellés",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "seen": {
                    "type": "boolean"
                  },
                  "flagged": {
                    "type": "boolean"
                  },
                  "labels": {
                    "type": "object",
                    "properties": {
                      "add": {
                        "maxItems": 20,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 64
                        }
                      },
                      "remove": {
                        "maxItems": 20,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 64
                        }
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMessagesId",
        "summary": "Met le message à la corbeille (ou le supprime)",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      }
    },
    "/messages/{id}/draft": {
      "get": {
        "operationId": "getMessagesIdDraft",
        "summary": "Corps d’un brouillon, pour le reprendre",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      }
    },
    "/messages/{id}/invitation": {
      "get": {
        "operationId": "getMessagesIdInvitation",
        "summary": "Invitation (iCalendar) portée par le message, analysée",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      }
    },
...
    "/messages/{id}/quote": {
      "get": {
        "operationId": "getMessagesIdQuote",
        "summary": "Citation pour répondre ou transférer",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ]
      }
    },
    "/messages/{id}/body": {
      "get": {
        "operationId": "getMessagesIdBody",
        "summary": "Corps HTML assaini (document pour iframe isolée)",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "text/html": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          },
          {
            "name": "images",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ]
            }
          },
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "light",
                "dark"
              ]
            }
          }
        ]
      }
    },
...
    "/image-proxy": {
      "get": {
        "operationId": "getImageProxy",
        "summary": "Image distante via le proxy (URL signée)",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "image/*": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 4096
            }
          },
          {
            "name": "exp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-plume_sid"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur (code stable, sans détail interne)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "required": [
                "error"
              ]
            }
          }
        }
      }
    }
  }
}